Replaced sample source file with new source files for the project
