+++ src/main.cpp
@@ -1,0 +1,129 @@
+#include "tracker.hpp"
+
+using namespace std;
+using namespace cv;
+
+namespace {
+const char* about = "Basic marker detection";
+
+//! [aruco_detect_markers_keys]
+const char* keys  =
+        "{v        |       | Input from video or image file, if ommited, input comes from camera }"
+        "{ci       | 0     | Camera id if input doesnt come from video (-v) }"
+        "{c        |       | Camera intrinsic parameters. Needed for camera pose }"
+        "{l        | 0.1   | Marker side length (in meters). Needed for correct scale in camera pose }";
+}
+//! [aruco_detect_markers_keys]
+
+int main(int argc, char *argv[]) {
+    CommandLineParser parser(argc, argv, keys);
+    parser.about(about);
+
+    if(argc < 2) {
+        parser.printMessage();
+        return 0;
+    }
+
+    bool estimatePose = parser.has("c");
+    float markerLength = parser.get<float>("l");
+
+    aruco::DetectorParameters detectorParams {cv::aruco::DetectorParameters()};
+
+    int camId = parser.get<int>("ci");
+
+    String video;
+    if(parser.has("v")) {
+        video = parser.get<String>("v");
+    }
+
+    if(!parser.check()) {
+        parser.printErrors();
+        return 0;
+    }
+
+    aruco::Dictionary dictionary = aruco::getPredefinedDictionary(cv::aruco::DICT_4X4_50);
+
+    Mat camMatrix, distCoeffs;
+    if(estimatePose) {
+        bool readOk = tracker::readCameraParameters(parser.get<string>("c"), camMatrix, distCoeffs);
+        if(!readOk) {
+            cerr << "Invalid camera parameters file" << endl;
+            return 0;
+        }
+    }
+
+    aruco::ArucoDetector detector(dictionary, detectorParams);
+    VideoCapture inputVideo;
+    int waitTime;
+    if(!video.empty()) {
+        inputVideo.open(video);
+        waitTime = 0;
+    } else {
+        inputVideo.open(camId);
+        waitTime = 10;
+    }
+
+    double totalTime = 0;
+    int totalIterations = 0;
+
+    // Set coordinate system
+    cv::Mat objPoints(4, 1, CV_32FC3);
+    objPoints.ptr<Vec3f>(0)[0] = Vec3f(-markerLength/2.f, markerLength/2.f, 0);
+    objPoints.ptr<Vec3f>(0)[1] = Vec3f(markerLength/2.f, markerLength/2.f, 0);
+    objPoints.ptr<Vec3f>(0)[2] = Vec3f(markerLength/2.f, -markerLength/2.f, 0);
+    objPoints.ptr<Vec3f>(0)[3] = Vec3f(-markerLength/2.f, -markerLength/2.f, 0);
+
+    std::cout << "Hit ESC key or Crtl + C to exit if a window opens." << std::endl;
+    while(inputVideo.grab()) {
+        Mat image, imageCopy, resizedImage;
+        inputVideo.retrieve(image);
+
+        resize(image, resizedImage, Size(426, 240));
+
+        double tick = (double)getTickCount();
+
+        vector< int > ids;
+        vector< vector< Point2f > > corners, rejected;
+
+        // detect markers and estimate pose
+        detector.detectMarkers(resizedImage, corners, ids, rejected);
+
+        size_t  nMarkers = corners.size();
+        vector<Vec3d> rvecs(nMarkers), tvecs(nMarkers);
+
+        if(estimatePose && !ids.empty()) {
+            // Calculate pose for each marker
+            for (size_t  i = 0; i < nMarkers; i++) {
+                solvePnP(objPoints, corners.at(i), camMatrix, distCoeffs, rvecs.at(i), tvecs.at(i));
+            }
+        }
+
+        double currentTime = ((double)getTickCount() - tick) / getTickFrequency();
+        totalTime += currentTime;
+        totalIterations++;
+        if(totalIterations % 30 == 0) {
+            cout << "Detection Time = " << currentTime * 1000 << " ms "
+                 << "(Mean = " << 1000 * totalTime / double(totalIterations) << " ms)" << endl;
+        }
+
+        // draw results
+        resizedImage.copyTo(imageCopy);
+        if(!ids.empty()) {
+            aruco::drawDetectedMarkers(imageCopy, corners, ids);
+
+            if(estimatePose) {
+                for(unsigned int i = 0; i < ids.size(); i++)
+                    cv::drawFrameAxes(imageCopy, camMatrix, distCoeffs, rvecs[i], tvecs[i], markerLength * 1.5f, 2);
+            }
+        }
+
+        if(!rejected.empty())
+            aruco::drawDetectedMarkers(imageCopy, rejected, noArray(), Scalar(100, 0, 255));
+
+        imshow("out", imageCopy);
+        char key = (char)waitKey(waitTime);
+        if(key == 27) break;
+    }
+
+    return 0;
+}
