+++ src/tracker.cpp
@@ -1,0 +1,6 @@
+#include "tracker.hpp"
+
+void tracker::trackLineFollower()
+{
+  
+}
